+++ src/app/shared/services/config.service.ts
@@ -2,10 +2,22 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable, OnDestroy } from '@angular/core';
 import { BehaviorSubject, Observable, Subject, take, takeUntil } from 'rxjs';
 
-import { handle } from '../utils';
+import { Fees, handle } from '../utils';
 import { WebSocketService } from './web-socket.service';
 
 import { ColorShades, PlatformConfig } from '@models/platformConfig';
+
+const emptyColorShades = {
+  color100: '',
+  color200: '',
+  color300: '',
+  color400: '',
+  color500: '',
+  color600: '',
+  color700: '',
+  color800: '',
+  color900: '',
+};
 
 export enum EXPENSE_TYPES {
   APORTE = 'Aporte',
@@ -258,14 +270,17 @@ export const DEFAULT_CONFIG = {
     hasStageName: true,
     hasImportants: true,
     hasMaterialList: true,
+    hasProvider: true,
+    hasMarketingPage: true,
+    units: [],
     businessFees: {
       support: {
-        organizationPercentage: '0,00',
-        nfPercentage: '0,00',
+        organizationPercentage: Fees.NORTAN_SUPPORT,
+        nfPercentage: Fees.NF_SUPPORT,
       },
       intermediation: {
-        nfPercentage: '0,00',
-        organizationPercentage: '0,00',
+        nfPercentage: Fees.NF_INTERMEDIATION,
+        organizationPercentage: Fees.NORTAN_INTERMEDIATION,
       },
     },
     codeAbbreviation: 'NINC',
@@ -292,19 +307,54 @@ export const DEFAULT_CONFIG = {
     hasExpertiseBySector: true,
   },
   socialConfig: {
+    showCompanyName: true,
+    address: '',
+    cnpj: '',
     youtubeLink: '',
     linkedinLink: '',
     instagramLink: '',
     glassfrogLink: '',
     gathertownLink: '',
     companyName: 'NINC',
+    logoDefault: {
+      url: '',
+      name: '',
+    },
+    logoWithoutName: {
+      url: '',
+      name: '',
+    },
+    logoWhite: {
+      url: '',
+      name: '',
+    },
+    logoWhiteWithoutName: {
+      url: '',
+      name: '',
+    },
+    colors: {
+      primary: emptyColorShades,
+      success: emptyColorShades,
+      info: emptyColorShades,
+      warning: emptyColorShades,
+      danger: emptyColorShades,
+    },
+    qrcodeURL: '',
   },
   modulesConfig: {
     hasPromotion: true,
     hasCourse: true,
   },
   oneDriveConfig: {
-    isActive: false,
+    isActive: true,
+    contracts: {
+      oneDriveId: '0',
+      folderId: '1',
+    },
+    providers: {
+      oneDriveId: '0',
+      folderId: '1',
+    },
   },
 };
